Debug Pgen counter and add geometric estimator

+++ generative_model_src/Counter.h
@@ -30,7 +30,7 @@ public:
 
 	virtual void initialize_counter(const Model_Parms& , const Model_marginals&) = 0;
 
-	virtual void count_scenario(double , double ,const std::string& , Seq_type_str_p_map& , const Seq_offsets_map& , const std::unordered_map<std::tuple<Event_type,Gene_class,Seq_side>, std::shared_ptr<Rec_Event>>&  , Mismatch_vectors_map& );
+	virtual void count_scenario(long double , double ,const std::string& , Seq_type_str_p_map& , const Seq_offsets_map& , const std::unordered_map<std::tuple<Event_type,Gene_class,Seq_side>, std::shared_ptr<Rec_Event>>&  , Mismatch_vectors_map& );
 
 	virtual void count_sequence(double , const Model_marginals& ,const Model_Parms&);
